+++ phase2/ATM_0354_phase2/UserHandler.java
@@ -80,7 +80,7 @@ public class UserHandler {
      * @param id The id of the account to withdraw from.
      * @return true if the withdrawal succeeds, false otherwise.
      */
-    boolean withdraw(User user, BigDecimal money, int id) throws MoneyTransferException {
+    boolean withdraw(User user, int money, int id) throws MoneyTransferException {
         for (Person person: this.users) {
             if (person == user) {
                 if (person instanceof User) {
